Removed "utils.js", added a table element for the game, added a reset button.
Removed the "utils.js" file which was going to be a class containing helper functions. This was removed because it'd just fragment the code without actually providing any modularization since this isn't a very complex project.

Added a table element that's going to be used to play Tic Tac Toe on. Added two spans to keep track of the player scores. Added a reset button. The JS/functionality for these will be added later.