Games of Tic-Tac-Toe can now be played.
This commit adds functionality to the table element (and, by extension, the application page as a whole). Full games of Tic-Tac-Toe can be played. User scores are saved by incrementing an integer in a span element by 1 each time. If the user refreshes the page, the scores are reset. As a result, there's a "play again" button that shows up at the end of each game to reset the board. The reset button next to the menu button resets the score. This allows the web app to be used offline if the user's browser has already loaded the page. Of course, any future Hue functionality I add won't work offline.