Added styling to the home page.
The home page now displays 4 cards that can be used to navigate to the different pages of the web app.

"utils.js" no longer in the head tag, so no more file not found errors.

Changed the titles of every page as the "Tic-Tac-Hue | Page" naming scheme was too long and would get cut off by the browser tab.

The "resize.css" file is now in every page's head tag.