Renamed some variables and modified the "light" function.
The light function now accepts two parameters, and instead gets the ID of the light from the user's browser's local storage. If no ID is found, then it uses 13 as the default ID (based on previous exercises on the DLE). The bulb's IP, ID, and the API key will, in the future, be able to be set in the settings page.